# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Modelo de API para transações da Saque e Pague. O parceiro interessado\
    \ deve implementar as operações desejadas para que a SeP consiga se comunicar\
    \ com o mesmo."
  version: "1.19.0"
  title: "Saque e Pague Transacional"
  contact:
    name: "Time Saque e Pague"
    url: "http://www.saqueepague.com.br"
host: "api-saqueepague.sensedia.com"
basePath: "/saqueepague/SepTransaction/1.19.0"
...
schemes:
- "https"
paths:
...
  /consultaPrePago:
    post:
      tags:
      - "Recarga Celular"
      description: "Operação de consulta dos possíveis produtos de recarga de cartã\
        o pré-pago. Operação relacionada a transação consultaPrePago."
      operationId: "consultaPrePagoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta produtos de recarga de cartão pré-pago."
        required: true
        schema:
          $ref: "#/definitions/consultaPrePagoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaPrePagoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "RecargaCelular"
  /consultaProdutosRecarga:
    post:
      tags:
      - "Recarga"
      description: "Operação de consulta dos possíveis produtos de recarga de transporte\
        \ para o cartão informado pelo cliente. Operação relacionada a transação consultaProdutosRecarga."
      operationId: "consultaProdutosRecargaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta produtos de recarga disponíveis para\
          \ um cartão."
        required: true
        schema:
          $ref: "#/definitions/consultaProdutosRecargaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaProdutosRecargaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "Recarga"
...
  /consultaValoresRecarga:
    post:
      tags:
      - "Recarga"
      description: "Operação de consulta dos produtos de recarga de cartão pré-pago\
        \ de transporte. Traz uma lista de todas as operadoras e seus respectivos\
        \ produtos/valores."
      operationId: "consultaValoresRecargaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de consulta valores de recarga de cartão."
        required: true
        schema:
          $ref: "#/definitions/consultaValoresRecargaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaValoresRecargaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "Recarga"
...
  /identificacao:
    post:
      tags:
      - "Saque com Cartão"
      - "Saldo com Cartão"
      - "Extrato com Cartão"
      - "Depósito com Cartão"
      - "Empréstimo com Cartão"
      - "Pagamento com Cartão"
      - "Transferência com Cartão"
      description: "Operação de identificação do cliente e informações adicionais\
        \ de sua conta."
      operationId: "identificacaoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de identificação."
        required: true
        schema:
          $ref: "#/definitions/identificacaoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/identificacaoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaqueComCarto"
...
  /recarga:
    post:
      tags:
      - "Recarga"
      description: "Operação de recarga de cartão pré-pago de transporte (bilhetagem)\
        \ em dinheiro."
      operationId: "recargaPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de operação de recarga de um cartão."
        required: true
        schema:
          $ref: "#/definitions/recargaReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/recargaResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "Recarga"
  /recargaConf:
    post:
      tags:
      - "Recarga"
      description: "Operação de confirmação de recarga de cartão pré-pago de transporte\
        \ (bilhetagem)."
      operationId: "recargaConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de recarga."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "Recarga"
  /recargaPrePago:
    post:
      tags:
      - "Recarga Celular"
      description: "Operação de recarga de celular pré-pago."
      operationId: "recargaPrePagoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição para recarga de celular pré-pago."
        required: true
        schema:
          $ref: "#/definitions/recargaPrePagoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/recargaPrePagoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "RecargaCelular"
  /recargaPrePagoConf:
    post:
      tags:
      - "Recarga Celular"
      description: "Confirmação de operação de recarga."
      operationId: "recargaPrePagoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de recarga de celular\
          \ pré-pago."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "RecargaCelular"
  /saldo:
    post:
      tags:
      - "Saldo com Cartão"
      description: "Operação de consulta do valor de saldo disponível em conta."
      operationId: "saldoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de operação de consulta de saldo."
        required: true
        schema:
          $ref: "#/definitions/saldoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/saldoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaldoComCarto"
  /saldoConf:
    post:
      tags:
      - "Saldo com Cartão"
      description: "Confirmação de operação de consulta de saldo."
      operationId: "saldoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de consulta de saldo."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaldoComCarto"
...
definitions:
...
  recargaPrePagoReq:
    type: "object"
    required:
    - "InfRecargaPrePago"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfRecargaPrePago:
        $ref: "#/definitions/infRecargaPrePagoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  recargaPrePagoResp:
    type: "object"
    required:
    - "InfRecargaPrePago"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfRecargaPrePago:
        $ref: "#/definitions/infRecargaPrePagoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfRecargaPrePago:
        idRecarga: "b1a5c4bd-a914-4506-8d1d-744a6b3a285c"
        bonus: "000000000400"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  consultaPrePagoReq:
    type: "object"
    required:
    - "InfConsultaPrePago"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfConsultaPrePago:
        $ref: "#/definitions/infConsultaPrePagoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaPrePagoResp:
    type: "object"
    required:
    - "InfConsultaPrePago"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfConsultaPrePago:
        $ref: "#/definitions/infConsultaPrePagoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaPrePago:
        produtos:
        - valorMaximo: "1500"
          idProduto: "CLARO"
          nomeOperadora: "CLARO"
          valorMinimo: "1000"
        - valorMaximo: "1500"
          idProduto: "CLARO"
          nomeOperadora: "CLARO"
          valorMinimo: "1000"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  consultaProdutosRecargaReq:
    type: "object"
    required:
    - "InfConsultaProdutosRecarga"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfConsultaProdutosRecarga:
        $ref: "#/definitions/infConsultaProdutosRecargaReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaProdutosRecargaResp:
    type: "object"
    required:
    - "InfConsultaProdutosRecarga"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfConsultaProdutosRecarga:
        $ref: "#/definitions/infConsultaProdutosRecargaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaProdutosRecarga:
        products:
        - name: "Produto de Recarga"
          id: "1"
        - name: "Produto de Recarga"
          id: "1"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  consultaValoresRecargaReq:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  consultaValoresRecargaResp:
    type: "object"
    required:
    - "InfConsultaValoresRecarga"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaValoresRecarga:
        $ref: "#/definitions/infConsultaValoresRecargaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      InfConsultaValoresRecarga:
        valores:
          issuers:
          - card_number_mask: "XXXXXXXXXXXXXX-X"
            card_number_length: 15
            name: "TEU"
            id: 1
            products:
            - min_value: "1000"
              fee: "200"
              values:
              - fee: "200"
                id: "1"
                value: "2000"
              - fee: "200"
                id: "1"
                value: "2000"
              name: "Bilhete Antecipado"
              id: "1"
              max_value: "9000"
            - min_value: "1000"
              fee: "200"
              values:
              - fee: "200"
                id: "1"
                value: "2000"
              - fee: "200"
                id: "1"
                value: "2000"
              name: "Bilhete Antecipado"
              id: "1"
              max_value: "9000"
          - card_number_mask: "XXXXXXXXXXXXXX-X"
            card_number_length: 15
            name: "TEU"
            id: 1
            products:
            - min_value: "1000"
              fee: "200"
              values:
              - fee: "200"
                id: "1"
                value: "2000"
              - fee: "200"
                id: "1"
                value: "2000"
              name: "Bilhete Antecipado"
              id: "1"
              max_value: "9000"
            - min_value: "1000"
              fee: "200"
              values:
              - fee: "200"
                id: "1"
                value: "2000"
              - fee: "200"
                id: "1"
                value: "2000"
              name: "Bilhete Antecipado"
              id: "1"
              max_value: "9000"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  criptoResp:
    type: "object"
    properties:
      hash:
        type: "string"
        description: "Hash de todos os campos que compõem a transação."
    example:
      hash: "hash"
...
  identificacaoReq:
    type: "object"
    required:
    - "Cartao"
    - "Cripto"
    - "InfIdentificacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfIdentificacao:
        $ref: "#/definitions/infIdentificacaoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  identificacaoResp:
    type: "object"
    required:
    - "InfIdentificacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfIdentificacao:
        $ref: "#/definitions/infIdentificacaoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfIdentificacao:
        pedirToken: "01"
        frase: "Informe o token gerado em seu app:"
        numAgencia: "4029"
        perguntas:
        - pergunta: "Qual é o dia de seu aniversário?"
          tamResposta: "2"
          id: "1"
        - pergunta: "Qual é o dia de seu aniversário?"
          tamResposta: "2"
          id: "1"
        idPositiva:
        - pos8: "4E-8G-1P"
          idTela: "1"
          pos4: "P1-G8-E4"
          pos5: "REG SOS"
          pos6: "EN/UQ/AM"
          pos7: "69 05 41"
          pos1: "SOS GER"
          pos2: "MA/QU/NE"
          pos3: "14 50 96"
        - pos8: "4E-8G-1P"
          idTela: "1"
          pos4: "P1-G8-E4"
          pos5: "REG SOS"
          pos6: "EN/UQ/AM"
          pos7: "69 05 41"
          pos1: "SOS GER"
          pos2: "MA/QU/NE"
          pos3: "14 50 96"
        numConta: "0082348296"
        transacValidas: "068101068010"
        saldo: "000000005000"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  infConsultaPrePagoReq:
    type: "object"
    required:
    - "ddd"
    - "telefone"
    properties:
      ddd:
        type: "string"
        example: "51"
        description: "Código da área do número de celular."
        pattern: "^\\d{2}$"
      telefone:
        type: "string"
        example: "982041311"
        description: "Número do telefone sem DDD."
        pattern: "^\\d+$"
    description: "Informações da requisição de consulta de produtos de recarga de\
      \ celular pré-pago."
...
  recargaReq:
    type: "object"
    required:
    - "InfRecarga"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfRecarga:
        $ref: "#/definitions/infRecargaReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  recargaResp:
    type: "object"
    required:
    - "InfRecarga"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfRecarga:
        $ref: "#/definitions/infRecargaResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfRecarga:
        recibo: "@             TEU BILHETE ANTECIPADO             @              \
          \                                  @ CARTAO: 036200005433591           \
          \             @ VALOR DA RECARGA: R$ 48,00                     @ TAXAS:\
          \ R$ 2,00                                 @ TOTAL PAGO: R$ 50,00       \
          \                    "
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  saldoReq:
    type: "object"
    required:
    - "InfSaldo"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfSaldo:
        $ref: "#/definitions/infSaldoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  saldoResp:
    type: "object"
    required:
    - "InfSaldo"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfSaldo:
        $ref: "#/definitions/infSaldoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfSaldo:
        recibo: "                 SALDO EM CONTA                @                036200005433591\
          \                @               13/10/2018  20:24                @    \
          \      BANCO: BANCO SAQUE E PAGUE           @          AGENCIA: 4029   \
          \                     @          CONTA: 0082348296                    @\
          \          VALOR NA DATA: R$ 50,00              "
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  transacConf:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoConf"
      Terminal:
        $ref: "#/definitions/terminal"
...